I merged more methods from TdsParserStateObject after #1520 It would maybe help to start working on fix the async bug.
