+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/TdsParserSessionPool.cs
@@ -182,7 +182,7 @@ namespace Microsoft.Data.SqlClient
                     // Session is good to re-use and our cache has space
                     SqlClientEventSource.Log.TryAdvancedTraceEvent("<sc.TdsParserSessionPool.PutSession|ADV> {0} keeping session {1} cachedCount={2}", ObjectID, session.ObjectID, _cachedCount);
 #if NETFRAMEWORK
-                    Debug.Assert(!session._pendingData, "pending data on a pooled session?");
+                    Debug.Assert(!session.HasPendingData, "pending data on a pooled session?");
 #else
                     Debug.Assert(!session.HasPendingData, "pending data on a pooled session?");
 #endif
